# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "开源、自托管轻量级发布平台 Ech0 的 API 文档",
        "title": "Ech0 API 文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6277",
    "basePath": "/api",
    "paths": {
...
        "/inbox": {
            "get": {
                "description": "根据分页条件获取系统收件箱",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "收件箱"
                ],
                "summary": "获取收件箱列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除所有收件箱消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "收件箱"
                ],
                "summary": "清空收件箱",
                "responses": {
                    "200": {
                        "description": "清空失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/inbox/{id}": {
            "delete": {
                "description": "根据 ID 删除收件箱消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "收件箱"
                ],
                "summary": "删除收件箱消息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "收件箱ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/inbox/{id}/read": {
            "put": {
                "description": "根据 ID 将消息标记为已读",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "收件箱"
                ],
                "summary": "标记消息为已读",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "收件箱ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "标记失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 状态码，非0时表示自定义HTTP业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "Data 响应数据，具体内容因接口而异"
                },
                "msg": {
                    "description": "Msg 返回信息，通常是状态描述",
                    "type": "string"
                }
            }
        },
...
    }
}